+++ Tests/Functional/V1/PreflightTest.php
@@ -7,14 +7,8 @@ declare(strict_types = 1);
  */
 namespace FindMyFriends\Functional\V1;
 
-use Elasticsearch;
 use FindMyFriends\Routing;
-use FindMyFriends\TestCase;
 use GuzzleHttp;
-use Hashids\Hashids;
-use Klapuch\Storage;
-use Klapuch\Uri;
-use PhpAmqpLib;
 use Psr\Http\Message;
 use Tester;
 use Tester\Assert;
@@ -22,8 +16,6 @@ use Tester\Assert;
 require __DIR__ . '/../../bootstrap.php';
 
 final class PreflightTest extends Tester\TestCase {
-	use TestCase\Redis;
-
 	/**
 	 * @dataProvider preflightHeaders
 	 */
@@ -52,26 +44,7 @@ final class PreflightTest extends Tester\TestCase {
 	}
 
 	private function endpoints(): array {
-		$matches = (new Routing\ApplicationRoutes(
-			new Uri\FakeUri(),
-			new class extends Storage\MetaPDO {
-				public function __construct() {
-				}
-			},
-			$this->redis,
-			Elasticsearch\ClientBuilder::create()->build(),
-			new PhpAmqpLib\Connection\AMQPLazyConnection(
-				'',
-				'',
-				'',
-				''
-			),
-			[
-				'demand' => ['hashid' => new Hashids()],
-				'evolution' => ['hashid' => new Hashids()],
-				'soulmate' => ['hashid' => new Hashids()],
-			]
-		))->matches();
+		$matches = (new Routing\TestApplicationRoutes())->matches();
 		return str_replace(' [OPTIONS]', '', preg_grep('~^v1/\w+ \[OPTIONS\]$~', array_keys($matches)));
 	}
