+++ Tests/Functional/V1/SortTest.php
@@ -7,17 +7,11 @@ declare(strict_types = 1);
  */
 namespace FindMyFriends\Functional\V1;
 
-use Elasticsearch;
 use FindMyFriends\Routing;
 use FindMyFriends\Schema;
-use FindMyFriends\TestCase;
 use FindMyFriends\V1\Demand;
 use FindMyFriends\V1\Demands;
 use GuzzleHttp;
-use Hashids\Hashids;
-use Klapuch\Storage;
-use Klapuch\Uri;
-use PhpAmqpLib;
 use Psr\Http\Message;
 use Tester;
 use Tester\Assert;
@@ -25,8 +19,6 @@ use Tester\Assert;
 require __DIR__ . '/../../bootstrap.php';
 
 final class SortTest extends Tester\TestCase {
-	use TestCase\Redis;
-
 	/**
 	 * @dataProvider sorts
 	 */
@@ -42,26 +34,7 @@ final class SortTest extends Tester\TestCase {
 				preg_grep(
 					'~sort=~',
 					array_keys(
-						(new Routing\ApplicationRoutes(
-							new Uri\FakeUri(),
-							new class extends Storage\MetaPDO {
-								public function __construct() {
-								}
-							},
-							$this->redis,
-							Elasticsearch\ClientBuilder::create()->build(),
-							new PhpAmqpLib\Connection\AMQPLazyConnection(
-								'',
-								'',
-								'',
-								''
-							),
-							[
-								'demand' => ['hashid' => new Hashids()],
-								'evolution' => ['hashid' => new Hashids()],
-								'soulmate' => ['hashid' => new Hashids()],
-							]
-						))->matches()
+						(new Routing\TestApplicationRoutes())->matches()
 					)
 				)
 			)
