+++ phpstan.neon
@@ -3,3 +3,4 @@ includes:
 parameters:
     ignoreErrors:
         - '~Call to an undefined method object::print\(\)~'
+        - '~AnonymousClass.+::__construct\(\) does not call parent constructor from Klapuch\\Storage\\MetaPDO~'
